requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.0
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
numpy==2.2.1
paramiko==3.5.0
pillow==11.3.0
pycparser==2.22
PyNaCl==1.5.0
sqlparse==0.5.3
tzdata==2025.2
